migration: fix auto load when enabled
