+++ extension/wxt.config.ts
@@ -64,6 +64,7 @@ export default defineConfig({
                         'fonts/*',
                         'asbplayer-locales/*',
                         'icon/image.png',
+                        'anime-page.js',
                         'netflix-page.js',
                         'youtube-page.js',
                         'stremio-page.js',
